cmd(core): move walkthroughs closer to usage
