+++ cmd/init.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"errors"
 	"os"
 	"path/filepath"
 
@@ -13,6 +14,36 @@ import (
 	"github.com/ubclaunchpad/inertia/local"
 	"github.com/ubclaunchpad/inertia/local/git"
 )
+
+var (
+	errInvalidBuildType     = errors.New("invalid build type")
+	errInvalidBuildFilePath = errors.New("invalid buildfile path")
+)
+
+// addProjectWalkthrough is the command line walkthrough that asks for details
+// about the project the user intends to deploy.
+func addProjectWalkthrough() (
+	buildType cfg.BuildType, buildFilePath string, err error,
+) {
+	resp, err := input.NewPrompt(nil).
+		PromptFromList("build type", []string{"docker-compose", "dockerfile"}).
+		GetString()
+	if err != nil {
+		return "", "", errInvalidBuildType
+	}
+	buildType, err = cfg.AsBuildType(resp)
+	if err != nil {
+		return "", "", err
+	}
+
+	buildFilePath, err = input.NewPrompt(nil).
+		Prompt(out.C("Please enter the path to your build configuration file:", out.CY)).
+		GetString()
+	if err != nil || buildFilePath == "" {
+		return "", "", errInvalidBuildFilePath
+	}
+	return
+}
 
 func attachInitCmd(inertia *core.Cmd) {
 	const (
@@ -128,7 +159,7 @@ See https://inertia.ubclaunchpad.com/#project-configuration for more details.`,
 			} else {
 				out.Println(":question: no build file detected")
 				var err error
-				buildType, buildFilePath, err = input.AddProjectWalkthrough()
+				buildType, buildFilePath, err = addProjectWalkthrough()
 				if err != nil {
 					out.Fatal(err)
 				}
